requirements.txt
# Screen capture and automation
pyautogui==0.9.54
pillow==10.0.0

# OCR (Optical Character Recognition)
pytesseract==0.3.10

# Computer vision
opencv-python==4.8.1.78

# GUI and system interaction
pynput==1.7.6

# Numpy for image processing
numpy==1.25.2

# Time and scheduling
schedule==1.2.0

# Configuration management
python-dotenv==1.0.0

# Logging
loguru==0.7.2
